Added database sql.

+++ bookstore/src/com/ymens/PrintAuthor.java
@@ -26,7 +26,7 @@ public class PrintAuthor {
                 author.description = resultSet.getString("description");
                 author.age = resultSet.getInt("age");
                 author.nationality = resultSet.getString("nationality");
-                author.cnp = resultSet.getInt("CNP");
+                author.cnp = resultSet.getLong("CNP");
             }
         }catch (SQLException e) {
             e.printStackTrace();
